requirements.txt
discord.py==2.4.0
pydantic==2.8.2
